requirements.txt
Flask==3.0.0
python-telegram-bot==20.7
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
pyotp==2.9.0
python-dateutil==2.8.2